fix change password empty update and add rating endpoint

+++ lib/app/data/model/category_destinations.dart
@@ -30,7 +30,7 @@ class Destination {
   String? destinationAddress;
   String? destinationImg;
   bool? destinationStauts;
-  int? destinationEvaluation;
+  double? destinationEvaluation;
 
   // bool? destinationFavorite;
